Refactored range test and moved it to test.verify
