+++ test/test_proc_format.py
@@ -33,7 +33,7 @@ class FormatTest():
         test_id = '{} format test'.format(proc_type)
 
         verify.equal_headers(test_id, FMT_INTEL_INT, FMT_INTEL_REAL, 'INT', 'REAL', False, True)
-        Base.data_is_equal(FMT_INTEL_INT, FMT_INTEL_REAL, 'INT', 'REAL')
+        verify.data_is_equal(FMT_INTEL_INT, FMT_INTEL_REAL, 'INT', 'REAL')
 
         print('INTEL FORMAT: OK')
 
@@ -48,7 +48,7 @@ class FormatTest():
         test_id = '{} format test'.format(proc_type)
 
         verify.equal_headers(test_id, FMT_DEC_INT, FMT_DEC_REAL, 'INT', 'REAL', False, True)
-        Base.data_is_equal(FMT_DEC_INT, FMT_DEC_REAL, 'INT', 'REAL')
+        verify.data_is_equal(FMT_DEC_INT, FMT_DEC_REAL, 'INT', 'REAL')
 
         print('DEC FORMAT: OK')
 
@@ -64,7 +64,7 @@ class FormatTest():
         test_id = '{} format test'.format(proc_type)
 
         verify.equal_headers(test_id, FMT_SGI_INT, FMT_SGI_REAL, 'INT', 'REAL', False, True)
-        Base.data_is_equal(FMT_SGI_INT, FMT_SGI_REAL, 'INT', 'REAL')
+        verify.data_is_equal(FMT_SGI_INT, FMT_SGI_REAL, 'INT', 'REAL')
 
         print('SGI FORMAT: OK')
 
@@ -81,9 +81,9 @@ class FormatTest():
         verify.equal_headers('INTEL-SGI INT format test', FMT_INTEL_INT, FMT_SGI_INT, 'INTEL', 'SGI', False, False)
         verify.equal_headers('DEC-SGI INT format test', FMT_DEC_INT, FMT_SGI_INT, 'DEC', 'SGI', False, False)
 
-        Base.data_is_equal(FMT_INTEL_INT, FMT_DEC_INT, 'INTEL', 'DEC')
-        Base.data_is_equal(FMT_INTEL_INT, FMT_SGI_INT, 'INTEL', 'SGI')
-        Base.data_is_equal(FMT_DEC_INT, FMT_SGI_INT, 'DEC', 'SGI')
+        verify.data_is_equal(FMT_INTEL_INT, FMT_DEC_INT, 'INTEL', 'DEC')
+        verify.data_is_equal(FMT_INTEL_INT, FMT_SGI_INT, 'INTEL', 'SGI')
+        verify.data_is_equal(FMT_DEC_INT, FMT_SGI_INT, 'DEC', 'SGI')
 
         print('INTEL-DEC-SGI INT FORMAT COMPARISON: OK')
 
@@ -101,9 +101,9 @@ class FormatTest():
         verify.equal_headers('INTEL-SGI INT format test', FMT_INTEL_REAL, FMT_SGI_REAL, 'INTEL', 'SGI', True, True)
         verify.equal_headers('DEC-SGI INT format test', FMT_DEC_REAL, FMT_SGI_REAL, 'DEC', 'SGI', True, True)
 
-        Base.data_is_equal(FMT_INTEL_REAL, FMT_DEC_REAL, 'INTEL', 'DEC')
-        Base.data_is_equal(FMT_INTEL_REAL, FMT_SGI_REAL, 'INTEL', 'SGI')
-        Base.data_is_equal(FMT_DEC_REAL, FMT_SGI_REAL, 'DEC', 'SGI')
+        verify.data_is_equal(FMT_INTEL_REAL, FMT_DEC_REAL, 'INTEL', 'DEC')
+        verify.data_is_equal(FMT_INTEL_REAL, FMT_SGI_REAL, 'INTEL', 'SGI')
+        verify.data_is_equal(FMT_DEC_REAL, FMT_SGI_REAL, 'DEC', 'SGI')
 
         print('INTEL-DEC-SGI REAL FORMAT COMPARISON: OK')
